+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.0.2
 
 - W.I.P. Planned to get character creation together as either a tGame command or in script process.
++ moved rjust, cjust, and title to ANSI class.
++ initial start to the menu[#] dictionary.
 
 ## v0.0.1